Fixed #4
Single row matrices now get parsed correctly instead of being parsed as
multi-row matrices, each with one element. Implemented using a flag
that keeps track of whether or not a vector is present inside matrix. If
not then the matrix must consist of one row.

Also modified the invalid matrix dimension error message to include
the expected dimensions of the matrix.