VectorOperator checks if number of arguments is valid
The logic to check whether or not the number of arguments is valid has not
yet been implemented. It will be implemented in the isValidArgumentLength
method of VectorOperator.

If the argument length is invalid an exception is thrown. The message of
which includes an expected usage String that must be provided by each
concrete subclass of VectorOperator.

This commit contains build errors in the VectorOperator subclasses.