Modified NodeFunctions transformation to a NodeNumber
If the argument length of the function is greater than zero (the function
has arguments) then an exception is thrown as the function cannot be
implicitly converted to a number without a value being provided for the
argument.

This modification improves the error messages recieved as otherwise
parse exceptions would be thrown when the function could not be evaluated.

Fixed #1