Fixed bug in pow operator
If the exponent was negative then the BigRational.pow method was not
being used meaning that the result was always a double. This lost
precision in the final answer. Now the correct method is used even if the
exponent is negative.

Modified expected result in Evaluator simple binary test case 33 due to
increased precision in the result